requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
Pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
transformers==4.44.2
torch==2.7.1
